Removed variables that depend on loop time length
• Removed rRateCount
• Removed SMA related variables
• rRate now calculated from gyro rather than SMA
• Shorter loop time (2000us to 250us)

+++ src/drone/DroneRadio.h
@@ -16,7 +16,7 @@ class DroneRadio {
   public:
     void init();
     void getInput();
-    void checkSignal(float loopTime, unsigned long currentTime);
+    void checkSignal(unsigned long loopTime, unsigned long currentTime);
     
     int timer; //Keeps track of loss of communication
